LPS-60972 The "copyLibs" task is specific to Liferay
